Input Type Validation
Made all functions that take input compatible with floats that are equal to ints. Fixed a few formatting issues.

Added type validation for __init__:
Checks wether p and n can be converted cleanly to int using %1. If they can't a TypeError is thrown. If %1 is zero then they are cast to int using int().
Added Type Validation for coordinates_from_distance:
If h%1 is not 0 a TypeError is thrown. If h%1 is 0 h is cast to int using int().
Standardized Error Messages in coordinates_from_distance:
Changed the ValueErrors thrown to the same formatting as the ValueErrors in __init__.
Added Type Validation for distance_from_coordinates:
If for any element in x element%1 is not 0 then a TypeError is thrown. Otherwise x is iterated through and each element is cast to int using int().
Formatting Correction in _binary_repr:
"width:int" ->"width: int"